facefusion-test/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
opencv-python==4.8.1.78
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
onnxruntime-gpu==1.16.3
numpy==1.24.4
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
psutil==5.9.6
